unicode fixes, compiler warning fixes, unsigned type unification, got rid of the shorts, log fixes...

+++ Graphics/macrodyngraphicsitem.h
@@ -40,8 +40,8 @@ public:
 
     void draw_mp_names(const QList<QString>&);// write multiple names in the window
 //    void draw_color_count();	// Job color_map
-    void set_axis(short, qreal, qreal);	// set max & min of axis
-    void get_axis(short, qreal*, qreal*); // get max & min of axis
+    void set_axis(int, qreal, qreal);	// set max & min of axis
+    void get_axis(int, qreal*, qreal*); // get max & min of axis
     void clear_window();		// clear output window
 
     void setPoint(const qreal&, const qreal&, const QColor&);
@@ -57,7 +57,7 @@ public:
     void dumpGraphics(const QString&) const; // dump output window
     QColor getBackgroundColor() const { return backgroundColor; }
     void setBackgroundColor(const QColor&);
-    QColor get_color(short);
+    QColor get_color(int);
 
     void clearColumn(qreal);	// clear a specified column of the
                     // output window
@@ -82,21 +82,21 @@ protected:
     xyRange *axis;
     QColor backgroundColor;
     QRect zoomRect;
-    quint32 lmargin;
-    quint32 rmargin;
-    quint32 upmargin;
-    quint32 lowmargin;
-    quint32 wid;
-    quint32 hig;
-    quint32 xinit, yinit, xpix_Min, ypix_Min, xpix_Max, ypix_Max;
-    quint32 right;
-    quint32 down;
+    uint lmargin;
+    uint rmargin;
+    uint upmargin;
+    uint lowmargin;
+    uint wid;
+    uint hig;
+    uint xinit, yinit, xpix_Min, ypix_Min, xpix_Max, ypix_Max;
+    uint right;
+    uint down;
     void getCoordinates(qreal, qreal, int&, int&);
     void draw_zoom_rect(int&, int&);
     double pixel_to_x(int);
     double pixel_to_y(int);
 
-    long colorCount[95];
+    qint64 colorCount[95];
 };
 
 #endif // MACRODYNGRAPHICSITEM_H
